requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
DateTime==4.9
dnspython==2.2.1
dominate==2.7.0
email-validator==1.3.0
exceptiongroup==1.1.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
MouseInfo==0.1.3
numpy==1.24.1
outcome==1.2.0
pandas==1.5.2
Pillow==9.3.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytweening==1.0.4
pytz==2022.7
requests==2.28.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.45
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
visitor==0.1.3
Werkzeug==2.2.2
wsproto==1.2.0
WTForms==3.0.1
zope.interface==5.5.2
